add docstrings BaseSerializer
